requirements.txt
aiogram==2.9.2
PyYAML==5.3.1
